+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 ## 2019-04-20
 ### Add
   - Add protobuf and generate message
+  - Generater grpc server
+  - add entrypoint
 
 ## 2019-04-16
 ### Add